+++ src/Curiosity.Migrations.PostgreSQL/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Improved cancellation support in `PostgresMigrationConnection`.
 - Improved parsing `PostgresException` in `MigrationActionHelper`.
+- Simplified fetching applied migrations
+
 
 ## [4.0.2] - 2023-04-10