+++ src/Curiosity.Migrations.SqlServer/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Improved cancellation token support.
+- Simplified fetching applied migrations
 
 ## [4.0.0-beta1] - 2023-05-01